List of commits I have applied
This should be everything that happened after they forked off
at last-ancestor and made their customizations, played back on
top of those customizations (i.e. a rebase of the rest of the 1.0.2
branch with pins.h and Configuration.h inserted way back in history,
plus manually resolving merge conflicts

In principle, this should have the all the same configuration settings
as the original firmware, pins should be defined correctly, and any
post-1.0.0 features in their default state.

Also to note, in theory firmware built from this should actually
properly drive my 3D printer...  To be determined if practice
is identical to theory.